requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==2.0.0
certifi==2022.9.24
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
distlib==0.3.6
Django==3.2
django-filter==22.1
django-pytest==0.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djoser==2.1.0
drf-base64==2.0
exceptiongroup==1.0.0
filelock==3.8.0
identify==2.5.8
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
nodeenv==1.7.0
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.10.0
